rename the http module to avoid weird import crashes
