+++ zavod/zavod/shed/zyte_api.py
@@ -5,7 +5,7 @@ from time import sleep
 
 from zavod import settings
 from zavod.context import Context
-from zavod.http import request_hash
+from zavod.runtime.http_ import request_hash
 
 
 class UnblockFailedException(RuntimeError):
